Cleanups and some debugging attempts for the write to register from ALU problems.
Oddly, the specific errors change when the base address of the control registers
in CPU32 change.  Some of the debugging seems to point to the connection from the
mux to the register file.  Though the path from the M0+ CPU through the mux to
the register file seems to work properly.